progress on moving things around.
