+++ src/Dogger.Tests/TestHelpers/TestCategories.cs
@@ -1,6 +1,6 @@
 ﻿namespace Dogger.Tests.TestHelpers
 {
-    class TestCategories
+    public class TestCategories
     {
         public const string UnitCategory = "Unit";
         public const string IntegrationCategory = "Integration";
