+++ src/Dogger.sln
@@ -23,6 +23,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{60474487
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dogger.Setup", "Dogger.Setup\Dogger.Setup.csproj", "{593773CC-EDEF-4198-B7CF-19549E94EA37}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dogger.Setup.Tests", "Dogger.Setup.Tests\Dogger.Setup.Tests.csproj", "{653399BD-B129-4FCC-B5FB-568B3F72648D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +43,10 @@ Global
 		{593773CC-EDEF-4198-B7CF-19549E94EA37}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{593773CC-EDEF-4198-B7CF-19549E94EA37}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{593773CC-EDEF-4198-B7CF-19549E94EA37}.Release|Any CPU.Build.0 = Release|Any CPU
+		{653399BD-B129-4FCC-B5FB-568B3F72648D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{653399BD-B129-4FCC-B5FB-568B3F72648D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{653399BD-B129-4FCC-B5FB-568B3F72648D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{653399BD-B129-4FCC-B5FB-568B3F72648D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
